move files to exmaples
